chore: move airtable service
